Adding env variable setup for Heroku deployment.
