+++ .gitignore
@@ -2,7 +2,7 @@
 .development-secrets.json
 node_modules
 docs
-dist/main.js
-dist/main.js.map
-dist/docs
+dist/*
+!dist/index.html
+!dist/assets/*
 **.swp
