+++ package.json
@@ -7,6 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack --config webpack.config.js",
     "start:dev": "webpack-dev-server --config webpack.development.config.js",
+    "start": "node index.js",
     "build:docs": "jsdoc src api -r -d dist/docs --readme README.md -c jsdoc.json",
     "lint": "eslint --ext .tsx --ext .ts src/ server.js"
   },
@@ -19,7 +20,7 @@
     "url": "https://github.com/Zero-Collateral/CeFi-Frontend/issues"
   },
   "engines": {
-    "node": "12.13.0"
+    "node": "14.6.0"
   },
   "homepage": "https://github.com/Zero-Collateral/CeFi-Frontend#readme",
   "dependencies": {
@@ -66,6 +67,7 @@
     "jsdoc": "^3.6.4",
     "parcel-bundler": "^1.12.4",
     "source-map-loader": "^0.2.4",
+    "terser": "3.14.1",
     "ts-loader": "^6.2.2",
     "typescript": "^3.8.3",
     "webpack": "^4.43.0",
